cutiuta-server/requirements.txt
apsw==3.8.10.1.post1
Beaker==1.6.4
BeautifulSoup==3.2.1
certifi==2016.2.28
Cython==0.22.1
Django==1.9.2
djangorestframework==3.4.6
docutils==0.12
gunicorn==19.4.5
lit==0.5.0.dev0
Mako==1.0.4
MarkupSafe==0.23
Pillow==3.1.1
pygobject==3.18.2
scons==2.4.1
vboxapi==1.0
virtualenv==13.1.2
wxPython==2.8.12.1
wxPython-common==2.8.12.1
